# apps/web/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "code": {
            "enum": [
              "not_found",
              "revision_conflict",
              "candidate_limit",
              "adopted_candidate",
              "candidate_archived",
              "candidate_provenance_immutable",
              "project_task_locked",
              "project_group_conflict",
              "protected_project",
              "project_has_successors",
              "data_integrity_error",
              "validation_error",
              "runtime_unavailable"
            ],
            "title": "Code",
            "type": "string"
          },
          "current_candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Candidate"
              },
              {
                "type": "null"
              }
            ]
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "Candidate": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inputs": {
            "$ref": "#/components/schemas/CandidateInputs"
          },
          "name": {
            "default": "候補",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "provenance": {
            "discriminator": {
              "mapping": {
                "copy": "#/components/schemas/CopySourceRef",
                "direct": "#/components/schemas/DirectSourceRef",
                "lineage": "#/components/schemas/LineageSourceRef",
                "manual": "#/components/schemas/ManualSourceRef",
                "screening": "#/components/schemas/ScreeningSourceRef",
                "snapshot": "#/components/schemas/SnapshotSourceRef"
              },
              "propertyName": "source_kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/ManualSourceRef"
              },
              {
                "$ref": "#/components/schemas/DirectSourceRef"
              },
              {
                "$ref": "#/components/schemas/LineageSourceRef"
              },
              {
                "$ref": "#/components/schemas/ScreeningSourceRef"
              },
              {
                "$ref": "#/components/schemas/SnapshotSourceRef"
              },
              {
                "$ref": "#/components/schemas/CopySourceRef"
              }
            ],
            "title": "Provenance"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "inputs",
          "id",
          "project_id",
          "revision",
          "created_at",
          "updated_at"
        ],
        "title": "Candidate",
        "type": "object"
      },
...
      "CopySourceRef": {
        "additionalProperties": false,
        "properties": {
          "source_kind": {
            "const": "copy",
            "title": "Source Kind",
            "type": "string"
          },
          "source_ref": {
            "$ref": "#/components/schemas/CopyReference"
          }
        },
        "required": [
          "source_kind",
          "source_ref"
        ],
        "title": "CopySourceRef",
        "type": "object"
      },
...
      "DataAsset": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "locator": {
            "minLength": 1,
            "title": "Locator",
            "type": "string"
          },
          "locator_kind": {
            "enum": [
              "managed",
              "bundled"
            ],
            "title": "Locator Kind",
            "type": "string"
          },
          "media_type": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Media Type",
            "type": "string"
          },
          "original_filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Original Filename",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "original_filename",
          "sha256",
          "media_type",
          "locator_kind",
          "locator",
          "id",
          "created_at"
        ],
        "title": "DataAsset",
        "type": "object"
      },
...
      "DataLibraryDataset": {
        "properties": {
          "data_asset": {
            "$ref": "#/components/schemas/DataAsset"
          },
          "dataset_revision": {
            "$ref": "#/components/schemas/DatasetRevision"
          },
          "dataset_views": {
            "items": {
              "$ref": "#/components/schemas/DatasetViewRevision"
            },
            "title": "Dataset Views",
            "type": "array"
          },
          "profile_revision": {
            "$ref": "#/components/schemas/ProfileRevision"
          },
          "supported_task_ids": {
            "items": {
              "type": "string"
            },
            "title": "Supported Task Ids",
            "type": "array"
          }
        },
        "required": [
          "dataset_revision",
          "data_asset",
          "profile_revision",
          "supported_task_ids"
        ],
        "title": "DataLibraryDataset",
        "type": "object"
      },
...
      "DatasetRevision": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "canonicalization_contract_digest": {
            "minLength": 1,
            "title": "Canonicalization Contract Digest",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data_asset_id": {
            "minLength": 1,
            "title": "Data Asset Id",
            "type": "string"
          },
          "dataset_digest": {
            "minLength": 1,
            "title": "Dataset Digest",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "profile_revision_id": {
            "minLength": 1,
            "title": "Profile Revision Id",
            "type": "string"
          }
        },
        "required": [
          "data_asset_id",
          "profile_revision_id",
          "canonicalization_contract_digest",
          "id",
          "dataset_digest",
          "created_at"
        ],
        "title": "DatasetRevision",
        "type": "object"
      },
      "DatasetRevisionUpdateInput": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          }
        },
        "required": [
          "archived"
        ],
        "title": "DatasetRevisionUpdateInput",
        "type": "object"
      },
      "DatasetViewMember": {
        "properties": {
          "cohort_key": {
            "default": "",
            "maxLength": 120,
            "title": "Cohort Key",
            "type": "string"
          },
          "cohort_label": {
            "default": "",
            "maxLength": 160,
            "title": "Cohort Label",
            "type": "string"
          },
          "dataset_revision_id": {
            "minLength": 1,
            "title": "Dataset Revision Id",
            "type": "string"
          },
          "dataset_view_revision_id": {
            "title": "Dataset View Revision Id",
            "type": "string"
          },
          "ordinal": {
            "minimum": 0.0,
            "title": "Ordinal",
            "type": "integer"
          },
          "provenance_json": {
            "additionalProperties": true,
            "title": "Provenance Json",
            "type": "object"
          }
        },
        "required": [
          "dataset_revision_id",
          "ordinal",
          "dataset_view_revision_id"
        ],
        "title": "DatasetViewMember",
        "type": "object"
      },
      "DatasetViewMemberInput": {
        "properties": {
          "cohort_key": {
            "default": "",
            "maxLength": 120,
            "title": "Cohort Key",
            "type": "string"
          },
          "cohort_label": {
            "default": "",
            "maxLength": 160,
            "title": "Cohort Label",
            "type": "string"
          },
          "dataset_revision_id": {
            "minLength": 1,
            "title": "Dataset Revision Id",
            "type": "string"
          },
          "ordinal": {
            "minimum": 0.0,
            "title": "Ordinal",
            "type": "integer"
          },
          "provenance_json": {
            "additionalProperties": true,
            "title": "Provenance Json",
            "type": "object"
          }
        },
        "required": [
          "dataset_revision_id",
          "ordinal"
        ],
        "title": "DatasetViewMemberInput",
        "type": "object"
      },
      "DatasetViewRevision": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "single",
              "cohort_comparison"
            ],
            "title": "Kind",
            "type": "string"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/DatasetViewMember"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "view_digest": {
            "minLength": 1,
            "title": "View Digest",
            "type": "string"
          },
          "view_id": {
            "title": "View Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "view_id",
          "revision",
          "name",
          "kind",
          "view_digest",
          "members",
          "created_at"
        ],
        "title": "DatasetViewRevision",
        "type": "object"
      },
      "DatasetViewRevisionCreateInput": {
        "properties": {
          "kind": {
            "enum": [
              "single",
              "cohort_comparison"
            ],
            "title": "Kind",
            "type": "string"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/DatasetViewMemberInput"
            },
            "minItems": 1,
            "title": "Members",
            "type": "array"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "view_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "View Id",
            "type": "string"
          }
        },
        "required": [
          "view_id",
          "revision",
          "name",
          "kind",
          "members"
        ],
        "title": "DatasetViewRevisionCreateInput",
        "type": "object"
      },
...
      "DirectSourceRef": {
        "additionalProperties": false,
        "properties": {
          "source_kind": {
            "const": "direct",
            "title": "Source Kind",
            "type": "string"
          },
          "source_ref": {
            "title": "Source Ref",
            "type": "null"
          }
        },
        "required": [
          "source_kind"
        ],
        "title": "DirectSourceRef",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "ModelPackageRef": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "locator": {
            "minLength": 1,
            "title": "Locator",
            "type": "string"
          },
          "manifest_digest": {
            "minLength": 1,
            "title": "Manifest Digest",
            "type": "string"
          },
          "manifest_json": {
            "additionalProperties": true,
            "title": "Manifest Json",
            "type": "object"
          },
          "package_id": {
            "minLength": 1,
            "title": "Package Id",
            "type": "string"
          },
          "task_contract_digest": {
            "minLength": 1,
            "title": "Task Contract Digest",
            "type": "string"
          },
          "task_id": {
            "minLength": 1,
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "package_id",
          "task_id",
          "task_contract_digest",
          "manifest_digest",
          "locator",
          "manifest_json",
          "id",
          "created_at"
        ],
        "title": "ModelPackageRef",
        "type": "object"
      },
      "ModelPackageRefUpdateInput": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          }
        },
        "required": [
          "archived"
        ],
        "title": "ModelPackageRefUpdateInput",
        "type": "object"
      },
...
      "ProfileRevision": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "canonical_contract_digest": {
            "minLength": 1,
            "title": "Canonical Contract Digest",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "effective_profile_json": {
            "additionalProperties": true,
            "title": "Effective Profile Json",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "profile_digest": {
            "minLength": 1,
            "title": "Profile Digest",
            "type": "string"
          },
          "profile_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Profile Id",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "profile_id",
          "revision",
          "name",
          "profile_digest",
          "canonical_contract_digest",
          "effective_profile_json",
          "id",
          "created_at"
        ],
        "title": "ProfileRevision",
        "type": "object"
      },
...
      "ProjectCreationOptions": {
        "properties": {
          "dataset_views": {
            "items": {
              "$ref": "#/components/schemas/DatasetViewRevision"
            },
            "title": "Dataset Views",
            "type": "array"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DataLibraryDataset"
            },
            "title": "Datasets",
            "type": "array"
          },
          "model_packages": {
            "items": {
              "$ref": "#/components/schemas/ModelPackageRef"
            },
            "title": "Model Packages",
            "type": "array"
          },
          "project_series": {
            "items": {
              "$ref": "#/components/schemas/ProjectSeries"
            },
            "title": "Project Series",
            "type": "array"
          },
          "task_contract_digests": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Task Contract Digests",
            "type": "object"
          }
        },
        "required": [
          "datasets",
          "dataset_views",
          "model_packages",
          "project_series",
          "task_contract_digests"
        ],
        "title": "ProjectCreationOptions",
        "type": "object"
      },
...
      "ProjectSeries": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectSeries",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Material Decision Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/data-library/datasets": {
      "get": {
        "operationId": "list_datasets_api_data_library_datasets_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DataLibraryDataset"
                  },
                  "title": "Response List Datasets Api Data Library Datasets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Datasets"
      }
    },
    "/api/data-library/datasets/{revision_id}": {
      "patch": {
        "operationId": "update_dataset_api_data_library_datasets__revision_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetRevisionUpdateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLibraryDataset"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Dataset"
      }
    },
    "/api/data-library/model-packages": {
      "get": {
        "operationId": "list_model_packages_api_data_library_model_packages_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelPackageRef"
                  },
                  "title": "Response List Model Packages Api Data Library Model Packages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Model Packages"
      }
    },
    "/api/data-library/model-packages/{reference_id}": {
      "patch": {
        "operationId": "update_model_package_api_data_library_model_packages__reference_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "reference_id",
            "required": true,
            "schema": {
              "title": "Reference Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelPackageRefUpdateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelPackageRef"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Model Package"
      }
    },
    "/api/data-library/views": {
      "get": {
        "operationId": "list_dataset_views_api_data_library_views_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetViewRevision"
                  },
                  "title": "Response List Dataset Views Api Data Library Views Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Dataset Views"
      },
      "post": {
        "operationId": "create_dataset_view_api_data_library_views_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetViewRevisionCreateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetViewRevision"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Dataset View"
      }
    },
...
    "/api/project-creation-options": {
      "get": {
        "operationId": "project_creation_options_api_project_creation_options_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreationOptions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Project Creation Options"
      }
    },
...
  }
}